Move state-related stuff to a separate package
